requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

Django==1.4.2
pinax-theme-bootstrap==2.2.2
#pinax-theme-bootstrap-account==1.0b2
#django-user-accounts==1.0b7
git+https://github.com/fay/pinax-theme-bootstrap-account
git+https://github.com/fay/django-user-accounts
django-forms-bootstrap==2.0.3.post1
metron==1.0
pinax-utils==1.0b1.dev3
eventlog==0.6
django-jsonfield==0.8.11
django-tagging
MySQL-python
PIL
requests
BeautifulSoup
django-tastypie==0.9.13-beta
kombu==2.4.3
celery==3.0.7
django-celery==3.0.11
django-celery-email==1.0.1
django-positions==0.4.3
git+https://github.com/fay/jieba
yuicompressor
django-pipeline==1.3.6
git+https://github.com/eldarion/idios.git
git+https://github.com/fay/pinax-theme-bootstrap
django-social-auth
